@@ -31,20 +31,21 @@
   "dependencies": {
     "@theia/application-package": "0.3.7",
     "bunyan": "^1.8.10",
-    "circular-dependency-plugin": "^2.0.0",
-    "copy-webpack-plugin": "^4.0.1",
+    "circular-dependency-plugin": "^4.0.0",
+    "copy-webpack-plugin": "^4.5.0",
     "css-loader": "^0.28.1",
-    "electron": "1.7.11",
+    "electron": "1.8.2-beta.5",
     "electron-rebuild": "^1.5.11",
-    "file-loader": "^0.11.1",
+    "file-loader": "^1.1.11",
     "font-awesome-webpack": "0.0.5-beta.2",
     "ignore-loader": "^0.1.2",
     "less": "^2.7.2",
     "source-map-loader": "^0.2.1",
     "source-map-support": "^0.4.18",
     "umd-compat-loader": "^2.1.1",
-    "url-loader": "^0.5.8",
-    "webpack": "^2.2.1"
+    "url-loader": "^1.0.1",
+    "webpack": "^4.0.0",
+    "webpack-cli": "^2.0.12"
   },
   "devDependencies": {
     "@theia/ext-scripts": "^0.2.0"
