@@ -5,31 +5,33 @@
   "main": "lib/common/index.js",
   "typings": "lib/common/index.d.ts",
   "dependencies": {
-    "@theia/application-package": "^0.3.7",
     "@phosphor/widgets": "^1.5.0",
+    "@theia/application-package": "^0.3.7",
     "@types/body-parser": "^1.16.4",
     "@types/bunyan": "^1.8.0",
     "@types/express": "^4.0.36",
+    "@types/lodash.debounce": "4.0.3",
+    "@types/lodash.throttle": "^4.1.3",
     "@types/perfect-scrollbar": "^1.3.0",
     "@types/ws": "^3.0.2",
     "@types/yargs": "^8.0.2",
-    "@types/lodash.throttle": "^4.1.3",
-    "lodash.throttle": "^4.1.1",
     "ajv": "^5.2.2",
     "body-parser": "^1.17.2",
     "bunyan": "^1.8.10",
-    "electron": "1.7.11",
+    "electron": "1.8.2-beta.5",
     "es6-promise": "^4.2.4",
     "express": "^4.15.3",
     "file-icons-js": "^1.0.3",
     "font-awesome": "^4.7.0",
     "inversify": "^4.2.0",
+    "lodash.debounce": "^4.0.8",
+    "lodash.throttle": "^4.1.1",
     "perfect-scrollbar": "^1.3.0",
     "reconnecting-websocket": "^3.0.7",
     "reflect-metadata": "^0.1.10",
-    "vscode-languageserver-types": "^3.4.0",
+    "vscode-languageserver-types": "^3.6.1",
     "vscode-uri": "^1.0.1",
-    "vscode-ws-jsonrpc": "0.0.1-alpha.5",
+    "vscode-ws-jsonrpc": "^0.0.2-1",
     "ws": "^3.0.0",
     "yargs": "^9.0.1"
   },
