@@ -13,7 +13,10 @@ import { isOSX } from '../common/os';
  * Since `M2+M3+<Key>` (Alt+Shift+<Key>) is reserved on MacOS X for writing special characters, such bindings are commonly
  * undefined for platform MacOS X and redefined as `M1+M3+<Key>`. The rule applies on the `M3+M2+<Key>` sequence.
  */
-export declare type Keystroke = { first?: Key, modifiers?: KeyModifier[] };
+export interface Keystroke {
+    readonly first?: Key;
+    readonly modifiers?: KeyModifier[];
+}
 
 export type KeySequence = KeyCode[];
 export namespace KeySequence {
@@ -82,9 +85,9 @@ export namespace KeySequence {
         return keyCodes;
     }
 
-    export function acceleratorFor(keysequence: KeySequence, separator: string = " ") {
+    export function acceleratorFor(keySequence: KeySequence, separator: string = " ") {
         const result: string[] = [];
-        for (const keyCode of keysequence) {
+        for (const keyCode of keySequence) {
             let keyCodeResult = "";
             let previous = false;
 
@@ -124,7 +127,20 @@ export namespace KeySequence {
                 if (previous) {
                     keyCodeResult += separator;
                 }
-                keyCodeResult += Key.getEasyKey(keyCode.key).easyString.toUpperCase();
+
+                const keyString = Key.getEasyKey(keyCode.key).easyString;
+
+                if (keyCode.key.keyCode >= Key.KEY_A.keyCode && keyCode.key.keyCode <= Key.KEY_Z.keyCode ||
+                    keyCode.key.keyCode >= Key.F1.keyCode && keyCode.key.keyCode <= Key.F24.keyCode) {
+                    /* We want to capitalize single letters and function keys */
+                    keyCodeResult += keyString.toUpperCase();
+                } else if (Key.getEasyKey(keyCode.key).easyString.length > 1) {
+                    /* We want to only capitalize the first letter */
+                    keyCodeResult += keyString.charAt(0).toUpperCase() + keyString.slice(1);
+                } else {
+                    /* Return the symbol as is */
+                    keyCodeResult += keyString;
+                }
             }
             result.push(keyCodeResult);
         }
@@ -310,8 +326,8 @@ export class KeyCode {
         throw new Error(`Cannot get key code from the keyboard event: ${event}.`);
     }
 
-    public static equals(keycode1: KeyCode, keycode2: KeyCode): boolean {
-        return JSON.stringify(keycode1) === JSON.stringify(keycode2);
+    public static equals(keyCode1: KeyCode, keyCode2: KeyCode): boolean {
+        return JSON.stringify(keyCode1) === JSON.stringify(keyCode2);
     }
 
     equals(event: KeyboardEvent | KeyCode): boolean {
@@ -365,6 +381,15 @@ export class KeyCode {
     }
 }
 
+export namespace KeyCode {
+
+    /**
+     * Determines a `true` of `false` value for the key code argument.
+     */
+    export type Predicate = (keyCode: KeyCode) => boolean;
+
+}
+
 export enum KeyModifier {
     /**
      * M1 is the COMMAND key on MacOS X, and the CTRL key on most other platforms.
@@ -412,8 +437,15 @@ export namespace KeyModifier {
     }
 }
 
-export declare type Key = { readonly code: string, readonly keyCode: number };
-export declare type EasyKey = { readonly keyCode: number, readonly easyString: string };
+export interface Key {
+    readonly code: string;
+    readonly keyCode: number;
+}
+
+export interface EasyKey {
+    readonly keyCode: number;
+    readonly easyString: string;
+}
 
 const CODE_TO_KEY: { [code: string]: Key } = {};
 const KEY_CODE_TO_KEY: { [keyCode: number]: Key } = {};
@@ -541,11 +573,12 @@ export namespace EasyKey {
 
 export namespace Key {
 
+    // tslint:disable-next-line:no-any
     export function isKey(arg: any): arg is Key {
         return !!arg && ('code' in arg) && ('keyCode' in arg);
     }
 
-    export function getKey(arg: string | number) {
+    export function getKey(arg: string | number): Key {
         if (typeof arg === "number") {
             return KEY_CODE_TO_KEY[arg] || {
                 code: 'unknown',
@@ -556,15 +589,19 @@ export namespace Key {
         }
     }
 
-    export function getEasyKey(key: Key) {
+    export function getEasyKey(key: Key): EasyKey {
         return KEY_CODE_TO_EASY[key.keyCode];
     }
 
-    export function isModifier(arg: string | number) {
+    export function isModifier(arg: string | number): boolean {
         if (typeof arg === "number") {
             return MODIFIERS.map(key => key.keyCode).indexOf(arg) > 0;
         }
         return MODIFIERS.map(key => key.code).indexOf(arg) > 0;
+    }
+
+    export function equals(key: Key, keyCode: KeyCode): boolean {
+        return !!keyCode.key && key.keyCode === keyCode.key.keyCode;
     }
 
     export const ENTER: Key = { code: "Enter", keyCode: 13 };
@@ -681,9 +718,9 @@ export namespace Key {
     });
     MODIFIERS.push(...[Key.ALT_LEFT, Key.ALT_RIGHT, Key.CONTROL_LEFT, Key.CONTROL_RIGHT, Key.O_S_LEFT, Key.O_S_RIGHT, Key.SHIFT_LEFT, Key.SHIFT_RIGHT]);
 
-    Object.keys(EasyKey).map(prop => Reflect.get(EasyKey, prop)).forEach(easykey => {
-        EASY_TO_KEY[easykey.easyString] = KEY_CODE_TO_KEY[easykey.keyCode];
-        KEY_CODE_TO_EASY[easykey.keyCode] = easykey;
+    Object.keys(EasyKey).map(prop => Reflect.get(EasyKey, prop)).forEach(easyKey => {
+        EASY_TO_KEY[easyKey.easyString] = KEY_CODE_TO_KEY[easyKey.keyCode];
+        KEY_CODE_TO_EASY[easyKey.keyCode] = easyKey;
     });
 })();
 
