@@ -5,11 +5,8 @@
  * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
  */
 
-import { inject, injectable } from 'inversify';
-import {
-    EditorManager, EditorDecorationsService, Range, Position, EditorDecoration,
-    EditorDecorationOptions, OverviewRulerLane
-} from '@theia/editor/lib/browser';
+import { injectable } from 'inversify';
+import { Range, Position, EditorDecoration, EditorDecorationOptions, OverviewRulerLane, TextEditor } from '@theia/editor/lib/browser';
 import { DirtyDiffUpdate } from './dirty-diff-manager';
 import { LineRange } from './diff-computer';
 
@@ -46,25 +43,24 @@ const ModifiedLineDecoration = <EditorDecorationOptions>{
 @injectable()
 export class DirtyDiffDecorator {
 
-    constructor(
-        @inject(EditorManager) protected readonly editorManager: EditorManager,
-        @inject(EditorDecorationsService) protected readonly editorDecorationsService: EditorDecorationsService,
-    ) { }
-
     applyDecorations(update: DirtyDiffUpdate): void {
         const modifications = update.modified.map(range => this.toDeltaDecoration(range, ModifiedLineDecoration));
         const additions = update.added.map(range => this.toDeltaDecoration(range, AddedLineDecoration));
         const removals = update.removed.map(line => this.toDeltaDecoration(line, RemovedLineDecoration));
         const decorations = [...modifications, ...additions, ...removals];
-        this.setDecorations(update.uri, decorations);
+        this.setDecorations(update.editor, decorations);
     }
 
     protected appliedDecorations = new Map<string, string[]>();
 
-    protected async setDecorations(uri: string, newDecorations: EditorDecoration[]) {
+    protected async setDecorations(editor: TextEditor, newDecorations: EditorDecoration[]) {
+        const uri = editor.uri.toString();
         const oldDecorations = this.appliedDecorations.get(uri) || [];
-        const decorationIds = await this.editorDecorationsService.deltaDecorations({ uri, oldDecorations, newDecorations });
-        this.appliedDecorations.set(uri, decorationIds || []);
+        if (oldDecorations.length === 0 && newDecorations.length === 0) {
+            return;
+        }
+        const decorationIds = await editor.deltaDecorations({ uri, oldDecorations, newDecorations });
+        this.appliedDecorations.set(uri, decorationIds);
     }
 
     protected toDeltaDecoration(from: LineRange | number, options: EditorDecorationOptions): EditorDecoration {
