@@ -8,7 +8,7 @@ import { injectable, inject } from 'inversify';
 import URI from '@theia/core/lib/common/uri';
 import { DisposableCollection, CommandRegistry, MenuModelRegistry } from '@theia/core';
 import { AbstractViewContribution, StatusBar, StatusBarAlignment, DiffUris } from '@theia/core/lib/browser';
-import { EditorManager, EditorWidget, EditorOpenerOptions, EditorContextMenu } from '@theia/editor/lib/browser';
+import { EditorManager, EditorWidget, EditorOpenerOptions, EditorContextMenu, EDITOR_CONTEXT_MENU } from '@theia/editor/lib/browser';
 import { GitFileChange } from '../common';
 import { GitWidget, GIT_WIDGET_CONTEXT_MENU } from './git-widget';
 import { GitRepositoryTracker } from './git-repository-tracker';
@@ -19,6 +19,8 @@ export const GIT_WIDGET_FACTORY_ID = 'git';
 const GIT_SELECTED_REPOSITORY = 'git-selected-repository';
 const GIT_REPOSITORY_STATUS = 'git-repository-status';
 const GIT_AHEAD_BEHIND = 'git-ahead-behind';
+
+export const EDITOR_CONTEXT_MENU_GIT = [...EDITOR_CONTEXT_MENU, '3_git'];
 
 export namespace GIT_COMMANDS {
     export const FETCH = {
@@ -80,7 +82,7 @@ export class GitViewContribution extends AbstractViewContribution<GitWidget> {
 
     onStart() {
         this.repositoryTracker.onDidChangeRepository(repository => {
-            if (repository) {
+            if (repository && this.hasMultipleRepositories()) {
                 const path = new URI(repository.localUri).path;
                 this.statusBar.setElement(GIT_SELECTED_REPOSITORY, {
                     text: `$(database) ${path.base}`,
@@ -161,7 +163,7 @@ export class GitViewContribution extends AbstractViewContribution<GitWidget> {
         });
         registry.registerCommand(GIT_COMMANDS.CHANGE_REPOSITORY, {
             execute: () => this.quickOpenService.changeRepository(),
-            isEnabled: () => this.repositoryTracker.allRepositories.length > 1
+            isEnabled: () => this.hasMultipleRepositories()
         });
         registry.registerCommand(GIT_COMMANDS.OPEN_FILE, {
             execute: () => this.openFile(),
@@ -215,4 +217,7 @@ export class GitViewContribution extends AbstractViewContribution<GitWidget> {
         return undefined;
     }
 
+    protected hasMultipleRepositories(): boolean {
+        return this.repositoryTracker.allRepositories.length > 1;
+    }
 }
