@@ -313,7 +313,7 @@ export interface CommitIdentity {
     /**
      * The date of the commit.
      */
-    readonly date: Date;
+    readonly timestamp: number;
 
     /**
      * The time-zone offset.
@@ -385,4 +385,15 @@ export declare enum GitError {
     ProtectedBranchRequiresReview = 34,
     ProtectedBranchForcePush = 35,
     PushWithPrivateEmail = 36
+}
+
+export interface GitFileBlame {
+    readonly uri: string;
+    readonly commits: Commit[];
+    readonly lines: CommitLine[];
+}
+
+export interface CommitLine {
+    readonly sha: string;
+    readonly line: number;
 }
