@@ -5,15 +5,17 @@
  * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
  */
 
-import { injectable, inject } from "inversify";
-import { h } from "@phosphor/virtualdom/lib";
-import { Message } from "@phosphor/messaging";
-import URI from "@theia/core/lib/common/uri";
+import { injectable, inject } from 'inversify';
+import { h } from '@phosphor/virtualdom/lib';
+import { Message } from '@phosphor/messaging';
+import URI from '@theia/core/lib/common/uri';
 import { SelectionService, CommandService } from '@theia/core/lib/common';
-import { ContextMenuRenderer, TreeProps, TreeModel, TreeNode, LabelProvider } from '@theia/core/lib/browser';
-import { FileTreeWidget, DirNode } from "@theia/filesystem/lib/browser";
+import { ContextMenuRenderer, TreeProps, TreeModel, TreeNode, LabelProvider, Widget, SelectableTreeNode } from '@theia/core/lib/browser';
+import { FileTreeWidget, DirNode } from '@theia/filesystem/lib/browser';
 import { WorkspaceService, WorkspaceCommands } from '@theia/workspace/lib/browser';
-import { FileNavigatorModel } from "./navigator-model";
+import { FileNavigatorModel } from './navigator-model';
+import { FileNavigatorSearch } from './navigator-search';
+import { SearchBox, SearchBoxProps, SearchBoxFactory } from './search-box';
 
 export const FILE_NAVIGATOR_ID = 'files';
 export const LABEL = 'Files';
@@ -22,6 +24,8 @@ export const CLASS = 'theia-Files';
 @injectable()
 export class FileNavigatorWidget extends FileTreeWidget {
 
+    protected readonly searchBox: SearchBox;
+
     constructor(
         @inject(TreeProps) readonly props: TreeProps,
         @inject(FileNavigatorModel) readonly model: FileNavigatorModel,
@@ -29,13 +33,31 @@ export class FileNavigatorWidget extends FileTreeWidget {
         @inject(CommandService) protected readonly commandService: CommandService,
         @inject(SelectionService) protected readonly selectionService: SelectionService,
         @inject(WorkspaceService) protected readonly workspaceService: WorkspaceService,
-        @inject(LabelProvider) protected readonly labelProvider: LabelProvider
+        @inject(LabelProvider) protected readonly labelProvider: LabelProvider,
+        @inject(FileNavigatorSearch) protected readonly navigatorSearch: FileNavigatorSearch,
+        @inject(SearchBoxFactory) protected readonly searchBoxFactory: SearchBoxFactory
     ) {
         super(props, model, contextMenuRenderer);
         this.id = FILE_NAVIGATOR_ID;
         this.title.label = LABEL;
         this.addClass(CLASS);
         this.initialize();
+        this.searchBox = searchBoxFactory(SearchBoxProps.DEFAULT);
+        this.toDispose.pushAll([
+            this.model.onExpansionChanged(() => this.searchBox.hide()),
+            this.navigatorSearch,
+            this.navigatorSearch.onFilteredNodesChanged(nodes => {
+                const node = nodes.find(SelectableTreeNode.is);
+                if (node) {
+                    this.model.selectNode(node);
+                }
+            }),
+            this.searchBox,
+            this.searchBox.onTextChange(data => this.navigatorSearch.filter(data)),
+            this.searchBox.onClose(data => this.navigatorSearch.filter(undefined)),
+            this.searchBox.onNext(() => this.model.selectNextNode()),
+            this.searchBox.onPrevious(() => this.model.selectPrevNode()),
+        ]);
     }
 
     protected initialize(): void {
@@ -80,6 +102,11 @@ export class FileNavigatorWidget extends FileTreeWidget {
         super.onAfterAttach(msg);
         this.addClipboardListener(this.node, 'copy', e => this.handleCopy(e));
         this.addClipboardListener(this.node, 'paste', e => this.handlePaste(e));
+        if (this.searchBox.isAttached) {
+            Widget.detach(this.searchBox);
+        }
+        Widget.attach(this.searchBox, this.node.parentElement!);
+        this.addKeyListener(this.node, this.searchBox.keyCodePredicate.bind(this.searchBox), this.searchBox.handle.bind(this.searchBox));
     }
 
     protected handleCopy(event: ClipboardEvent): void {
