@@ -5,12 +5,8 @@
  * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
  */
 import * as chai from 'chai';
-import 'mocha';
-import * as chaiAsPromised from 'chai-as-promised';
 import { testContainer } from './test/inversify.spec-config';
 import { IShellTerminalServer } from '../common/shell-terminal-protocol';
-
-chai.use(chaiAsPromised);
 
 /**
  * Globals
@@ -23,8 +19,9 @@ describe('ShellServer', function () {
     this.timeout(5000);
     const shellTerminalServer = testContainer.get<IShellTerminalServer>(IShellTerminalServer);
 
-    it('test shell terminal create', function () {
+    it('test shell terminal create', async function () {
         const createResult = shellTerminalServer.create({});
-        return expect(createResult).to.be.eventually.greaterThan(-1);
+
+        expect(await createResult).to.be.greaterThan(-1);
     });
 });
