@@ -5,9 +5,6 @@
  * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
  */
 
-import * as chai from 'chai';
-import 'mocha';
-import * as chaiAsPromised from 'chai-as-promised';
 import { testContainer } from './test/inversify.spec-config';
 import { BackendApplication } from '@theia/core/lib/node/backend-application';
 import { IShellTerminalServer } from '../common/shell-terminal-protocol';
@@ -15,14 +12,6 @@ import * as ws from 'ws';
 import * as http from 'http';
 import * as https from 'https';
 import { terminalsPath } from '../common/terminal-protocol';
-
-chai.use(chaiAsPromised);
-
-/**
- * Globals
- */
-
-const expect = chai.expect;
 
 describe('Terminal Backend Contribution', function () {
 
@@ -48,6 +37,7 @@ describe('Terminal Backend Contribution', function () {
                 reject(error);
             });
         });
-        return expect(p).to.be.eventually.fulfilled;
+
+        await p;
     });
 });
